+++ DirectX Engine/MathHelper.h
@@ -3,13 +3,13 @@
 class MathHelper
 {
 public:
-	static constexpr float PI2_f =  3.141592653589793238 * 2;
-	static constexpr float PI_f =      3.141592653589793238;
-	static constexpr float PI_2_f =    3.141592653589793238 / 2;
-	static constexpr float PI_4_f = 3.141592653589793238 / 4;
+	static constexpr float PI2_f =  3.141592653589793238f * 2;
+	static constexpr float PI_f =      3.141592653589793238f;
+	static constexpr float PI_2_f =    3.141592653589793238f / 2;
+	static constexpr float PI_4_f = 3.141592653589793238f / 4;
 
-	static constexpr double PI2 = 3.141592653589793238 * 2;
-	static constexpr double PI = 3.141592653589793238;
-	static constexpr double PI_2 = 3.141592653589793238 / 2;
-	static constexpr double PI_4 = 3.141592653589793238 / 4;
+	static constexpr float PI2 = 3.141592653589793238f * 2;
+	static constexpr float PI = 3.141592653589793238f;
+	static constexpr float PI_2 = 3.141592653589793238f / 2;
+	static constexpr float PI_4 = 3.141592653589793238f / 4;
 };
